logout hotfix etc pt.2
